This update contains the underlying database of the mainwebapp.
The following instructions have been added to 'resources/install-notes.txt'
::Database::
The mainwebapp requires account information stored in an underlying database.
Until there is an automated way of the application creating the database for us, it needs to be done with the following steps:
>Run the following java-app: org.duraspace.mainwebapp.domain.repo.db.MainDatabaseUtil <location-of-new-DB>

To ensure that the classpath is setup properly, it is probably easiest to run this from within eclipse.

>Now the mainwebapp needs to know where this newly created database is.
 Spring injects the dataSource which it receives with a jndi lookup from the container.
 To let the container know where the database is, set the same location that you used above (<location-of-new-DB>) in the following file:
>mainwebapp/src/main/webapp/META-INF/context.xml
>Currently the line is set as:
 url="jdbc:derby:/opt/derby/duraspaceDB"
>Change it to:
 url="jdbc:derby:<location-of-new-DB>"




git-svn-id: https://svn.duraspace.org/duracloud-priv@45 3c0c4179-e91d-42e7-a082-8e2792ce3676