Implementation of optional Content-Type and Content-Length getBlob hints:
Content-Type: If specified, this will be used as the mime type given to ContentStore for writes. If unspecified, application/octet-stream will be used.

Content-Length: If specified, this will be used as the content size passed to ContentStore for writes. This must be accurate if given, whereas the estimatedSize argument for Blob.openOutputStream is not required to be as per the contract defined by Akubra (in fact, estimatedSize is not used by this implementation because it's not useful if it's innacurate).

Note: Whenever moveTo is called, these values will always be passed to ContentStore. First, if present, the values in the hints for the source blob are used. If not present, the stored values for the original blob are used.


git-svn-id: https://svn.duraspace.org/duracloud-priv@456 3c0c4179-e91d-42e7-a082-8e2792ce3676