Adds a new database utility which can: 1. Retrieve all data from all database domains and store it locally in an easy-to-read XML format 2. Push updated or new info into any of the database domains 3. Clear out all of the content in all database domains
The initial impetus for this tool was to provide a way to easily get information in and out of the AMA database without needing to use a UI (which may not yet exist).
To use the tool, build the account-management-db-util project, step into the target directory, and run java -jar am-db-util-driver.jar, which will provide a quick overview of the required parameters. For more information, see the description of the DbUtil class.

This commit also:
- finishes adding in the InstanceRepo
- pushes some (web-specific) dependencies down from the top level pom into the AMA app pom
- removes the unnecessary Role field in UserInvitation
- adds a few other minor tweaks

git-svn-id: https://svn.duraspace.org/duracloud-priv/trunk@906 3c0c4179-e91d-42e7-a082-8e2792ce3676