This update resolves https://jira.duraspace.org/browse/DURACLOUDPRIV-20 . It sends email notifications on the account creation event to both the system admins and the user who created the account. A new, repeatable, config element is added to the ama-init.properties which specifies the email addresses of the AMA admins (i.e. Carissa). See updates to 'init.properties' for details.
Also, the class AmaEndpoint has been created to hold the static host/port/context config of the running AMA.
This is particularly helpful in sending emails that need a link back to the AMA.

Also, this check-in upgrades EasyMock to version 3.0.
The primary difference from our immediate perspective is that the functionality of the existing package:
  org.easymock.classextension
has been rolled into the standard package:
  org.easymock

So any place that you notice EasyMock as deprecated, remove the 'classextension' portion of the import path.



git-svn-id: https://svn.duraspace.org/duracloud-priv/trunk@946 3c0c4179-e91d-42e7-a082-8e2792ce3676