This update accomplishes two related tasks in preparation for rolling out the improved root console and account creation wizard. 1. It upgrades the html doc type to use the html 5 (<!DOCTYPE html>) and we are producing text/html rather than application/xhtml+xml.  In order to make this work I've also made sure to put place holders all empty script, div, i, and other tags, to keep the browsers happy. 2. In the process of upgrading there were abundant opportunities to remove obsolete code and tidy up messy code.
I've refrained from making any visible changes to the layout as well as big rewrites to the jspx pages  - things should look the same and all functions should continue to work as before. There is a second root console link which is not connected to anything yet.  My strategy is to rework the management console and then apply the new code organization to the rest of the site in stages.




git-svn-id: https://svn.duraspace.org/duracloud-priv/trunk@1226 3c0c4179-e91d-42e7-a082-8e2792ce3676