DURACLOUD-1227: Adds a new S3 Swift storage provider option to the Management Console (PR #24)
* Adds Swift S3 as a storage option

To accommodate changes made in duradmin and durastore, we have added an option to the UI to select Swift S3 as a storage provider. Selecting this will give the user the option to input their URL endpoint and the Signer Type they are using.

* Adds Amazon S3 and Swift S3 as secondary providers

Changes Amazon S3 to no longer be a primary-only storage option, so that it can be added to the list of valid secondary providers. Any provider type can now be primary.