Switch constants for Enums and rename variables
We now use Enums instead of constants to make it easier to see what the
valid values for each "type" are.

We have also changed any instance of auditQueueType to queueType, to
reduce the number of "types".

Any instance of "AWS" has been changed to the more specific AWS service
in question.